Media tags implementation

+++ backend/src/LinkModules/tagging/tagging.service.ts
@@ -1,11 +1,11 @@
 import { Injectable, InternalServerErrorException } from '@nestjs/common';
-import { CreateTaggingDto } from './dto/create-tagging.dto';
-import { UpdateTaggingDto } from './dto/update-tagging.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Tagging } from './entities/tagging.entity';
-import { Repository } from 'typeorm';
+import { DeleteResult, Repository } from 'typeorm';
 import { TagService } from '../../BaseEntities/tag/tag.service';
 import { CustomLogger } from '../../utils/Logger/CustomLogger.service';
+import { UserGroupService } from '../../BaseEntities/user-group/user-group.service';
+import { ObjectTypes } from '../../enum/ObjectTypes';
 
 @Injectable()
 export class TaggingService {
@@ -15,21 +15,38 @@ export class TaggingService {
     @InjectRepository(Tagging)
     private readonly taggingRepository: Repository<Tagging>,
     private readonly tagsService: TagService,
+    private readonly userGroupService: UserGroupService,
   ) {}
 
   async assignTagToObject(
-    tagName: string,
-    objectType: string,
+    tagTitle: string,
     objectId: number,
-  ): Promise<void> {
+    userPersonalGroupId: number,
+    objectTypes: ObjectTypes,
+  ): Promise<Tagging> {
     try {
-      const tag = await this.tagsService.createTag(tagName, true);
+      const userPersonalGroup =
+        await this.userGroupService.findUserPersonalGroup(userPersonalGroupId);
+      let tag = await this.tagsService.findTagByName(tagTitle);
+      if (!tag) {
+        tag = await this.tagsService.createTag({ title: tagTitle });
+      }
       const tagging = this.taggingRepository.create({
-        tagId: tag.id,
-        objectType,
+        tag: tag,
         objectId,
+        objectTypes,
       });
-      await this.taggingRepository.save(tagging);
+      await this.taggingRepository.upsert(tagging, {
+        conflictPaths: ['tag', 'objectTypes', 'objectId', 'user'],
+      });
+
+      return this.taggingRepository.findOne({
+        where: {
+          tag: tag,
+          objectId: objectId,
+          objectTypes: objectTypes,
+        },
+      });
     } catch (error) {
       this.logger.error(error.message, error.stack);
       throw new InternalServerErrorException(
@@ -39,13 +56,10 @@ export class TaggingService {
     }
   }
 
-  async getTagsForObject(
-    objectType: string,
-    objectId: number,
-  ): Promise<Tagging[]> {
+  async getTagsForObject(objectId: number): Promise<Tagging[]> {
     try {
       return this.taggingRepository.find({
-        where: { objectType, objectId },
+        where: { objectId },
         relations: ['tag'],
       });
     } catch (error) {
@@ -58,14 +72,24 @@ export class TaggingService {
   }
 
   async removeTagFromObject(
-    tagName: string,
-    objectType: string,
+    tagTitle: string,
+    objectTypes: ObjectTypes,
     objectId: number,
-  ): Promise<void> {
-    const tag = await this.tagsService.createTag(tagName); // find or create tag if it doesn't exist
-    await this.taggingRepository.delete({
-      tagId: tag.id,
-      objectType,
+  ): Promise<DeleteResult> {
+    const taggingForObject = await this.getTagsForObject(objectId);
+    const tagToRemove = taggingForObject.find(
+      (tagging) =>
+        tagging.tag.title === tagTitle && tagging.objectTypes === objectTypes,
+    );
+    if (!tagToRemove) {
+      throw new Error(
+        `Tag with title "${tagTitle}" not found for the specified object.`,
+      );
+    }
+
+    return await this.taggingRepository.delete({
+      tagId: tagToRemove.tagId,
+      objectTypes: tagToRemove.objectTypes,
       objectId,
     });
   }
